данные для поиска карты перенесны в json
