--- files/bins.txt
@@ -1,15 +1,0 @@
-510126
-519778
-521178
-522828
-523701
-530331
-530827
-548673
-548674
-552175
-555928
-555933
-555949
-555957
-555921
