--- files/hash.txt
@@ -1,1 +1,0 @@
-4006234246b4fd2b2833d740927ab20465afad862c74b1a88ec0869bde5c836c
