--- files/last_numbers.txt
@@ -1,1 +1,0 @@
-0254
