--- files/settings.json
@@ -1,1 +1,0 @@
-{"hash_file": "files/hash.txt", "bins_file": "files/bins.txt", "last_numbers_file": "files/last_numbers.txt"}
