+++ oib-laba-4.py
@@ -6,14 +6,17 @@ from window import Window
 import json
 
 SETTINGS = {
-    'hash_file': 'files/hash.txt',
-    'bins_file': 'files/bins.txt',
-    'last_numbers_file': 'files/last_numbers.txt',
+    'hash': '4006234246b4fd2b2833d740927ab20465afad862c74b1a88ec0869bde5c836c',
+    'bins': ('510126', '519778', '521178', '522828', '523701',
+        '530331', '530827', '548673', '548674', '552175',
+        '555928', '555933', '555949', '555957', '555921'
+        ),
+    'last_numbers': '0254',
 }
 
 if __name__ == '__main__':
     logging.basicConfig(level=logging.INFO, filename='logs.log', filemode='w')
-    with open('files/settings.json', 'w') as fp:
+    with open('settings.json', 'w') as fp:
         json.dump(SETTINGS, fp)
 
     app = QApplication(sys.argv)
