+++ settings.json
@@ -1,0 +1,1 @@
+{"hash": "4006234246b4fd2b2833d740927ab20465afad862c74b1a88ec0869bde5c836c", "bins": ["510126", "519778", "521178", "522828", "523701", "530331", "530827", "548673", "548674", "552175", "555928", "555933", "555949", "555957", "555921"], "last_numbers": "0254"}
